+++ classes/plugin.php
@@ -42,7 +42,9 @@ class block_quickmail_plugin {
      * @throws required_capability_exception
      */
     public static function require_user_can_send($sendtype, $user, $context, $page) {
-        if (!self::user_can_send($sendtype, $user, $context, $page, false)) {
+        // if (!self::user_can_send($sendtype, $user, $context, $page, false)) {
+        $includestudentaccess = -1 !== (int) get_config('moodle', 'block_quickmail_allowstudents');
+        if (!self::user_can_send($sendtype, $user, $context, $page, $includestudentaccess)) {
             $capability = $sendtype == 'broadcast' ? 'myaddinstance' : 'cansend';
 
             throw new required_capability_exception($context, 'block/quickmail:' . $capability, 'nopermissions', '');
@@ -139,11 +141,12 @@ class block_quickmail_plugin {
             return false;
         }
 
-        if (self::check_frozen_context($user, $context, $page)) {
+        $usercapa = self::user_has_capability('cansend', $user, $context);
+        if ($usercapa) {
             return true;
         }
 
-        if (self::user_has_capability('cansend', $user, $context)) {
+        if (self::check_frozen_context($user, $context, $page) && $usercapa) {
             return true;
         }